Btn moved to comp folder
