+++ src/components/Calendar/Calendar.module.scss
@@ -4,7 +4,8 @@
 .wrapper {
   border: 1px solid var(--porcelain);
   box-shadow: 0 4px 8px rgba(24, 27, 43, 0.04),
-    0 8px 16px rgba(24, 27, 43, 0.05), 0 12px 32px rgba(24, 27, 43, 0.05);
+    0 8px 16px rgba(24, 27, 43, 0.05),
+    0 12px 32px rgba(24, 27, 43, 0.05);
   border-radius: $border-radius;
   width: 246px;
   background: var(--white);
