+++ src/ducks/SANCharts/ChartSettings.js
@@ -1,7 +1,7 @@
 import React from 'react'
 import Checkboxes from '@santiment-network/ui/Checkboxes'
 import Selector from '@santiment-network/ui/Selector/Selector'
-import CalendarBtn from './CalendarBtn'
+import CalendarBtn from '../../components/Calendar/CalendarBtn'
 import SearchProjects from '../../components/Search/SearchProjects'
 import ShareModalTrigger from '../../components/Share/ShareModalTrigger'
 import styles from './ChartPage.module.scss'
